AI Providers: Back away from pre-initialized model attendants paradigm.
(There are places where we do want to recover the model object directly
from the attendant without reconstructing the object. Also, the coupling
involved with attendant classes adds complexity to both the interface
and implementation. The overhead for producing attendants on-the-fly is
miminum and does allow for response to dynamically-updated model
objects.)

* Remove 'ConverserAttendants' and 'ConverserAttendantClasses' classes.

* Convert 'ConverserSerdeProcessor' to initialize from a model rather
  than model address and attributes.