GUI: Update summarization feature.
Do not deactivate documents, invocations, and invocation results in
history above summarization request.

Originally, the summarization feature was designed in the era of 4K and
8K context windows as a means of usefully extending a conversation. Now,
it is useful as a cost optimization tool, since passing lots of tokens
is expensive. Invocation (function calling, tool use) was not available
with any model when the feature was developed. This change accounts for
invocations as well as an old TODO item to preserve documents from
semantic queries against vector databases (now called
"Retrieval-Augmented Generation (RAG)"). This will also be useful for
"persistent project files" when we implement the project concept as we
can reuse the 'Document' role for those.